Delete old sharing APIs
