+++ doc/sharing.rst
@@ -36,59 +36,3 @@ shared memory for parallel processing.
 
 .. autoclass:: PersistedModel
     :members:
-
-
-Model Store API
----------------
-
-Model stores handle persisting models into shared memory, cleaning up shared memory, and
-making objects available to other classes.
-
-LensKit users and algorithm implementers will generally not need to use this code themselves,
-unlessthey are implementing their own batch processing logic.
-
-.. autofunction:: get_store
-
-.. autoclass:: BaseModelStore
-    :members:
-    :show-inheritance:
-
-.. autoclass:: BaseModelClient
-    :members:
-    :show-inheritance:
-
-.. autoclass:: SharedObject
-    :members:
-    :show-inheritance:
-
-Model Store Implementations
----------------------------
-
-We provide several model store implementations.
-
-Memory Mapping
-~~~~~~~~~~~~~~
-
-.. py:module:: lenskit.sharing.file
-
-The memory-mapped-file store works on any supported platform and Python version.  It uses
-BinPickle's memory-mapped file support extension to store models on disk and use their
-storage to back memory-mapped views of major data structures.
-
-.. autoclass:: FileModelStore
-    :show-inheritance:
-.. autoclass:: FileClient
-    :show-inheritance:
-
-Shared Memory
-~~~~~~~~~~~~~
-
-.. py:module:: lenskit.sharing.sharedmem
-
-This store uses Python 3.8's :py:mod:`multiprocessing.shared_memory` module, along with out-of-band
-buffer support in Pickle Protcol 5, to pass model data through shared memory.
-
-.. autoclass:: SHMModelStore
-    :show-inheritance:
-.. autoclass:: SHMClient
-    :show-inheritance:
